pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "msatutil"
description = "Utility codes to read and plot from MethaneSAT/AIR files"
authors = [
    { name = "Sebastien Roche", email = "sroche@g.harvard.edu" }
]
license = { file = "LICENSE.md" }
requires-python = ">=3.9,<3.12"
urls."Homepage" = "https://github.com/rocheseb/msatutil"
classifiers = [
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dynamic = ["version"]
dependencies = [
    "dask[dataframe]",
    "netcdf4",
    "matplotlib",
    "pandas",
    "scipy",
    "tqdm",
    "google-cloud-storage",
    "pyogrio",
]

[project.optional-dependencies]
notebooks = [
    "notebook",
    "panel",
    "holoviews",
    "cartopy",
    "geoviews",
    "geopandas",
    "datashader"
]
gcsfs = [
    "fsspec",
    "gcsfs"
]

[tool.setuptools]
include-package-data = true
zip-safe = false
license-files = ["LICENSE.md"]

[tool.setuptools.packages]
find = {namespaces = false}

[project.entry-points."console_scripts"]
mairhtml = "msatutil.mair_geoviews:main"
mairls = "msatutil.mair_ls:main"
msattargets = "msatutil.msat_targets:main"

[tool.setuptools_scm]
write_to = "msatutil/_version.py"
version_scheme = "post-release"
